Controller for Cliente and Meeting, changed type for numbers in models and created README file

+++ src/main/java/com/alexis/meetingback/models/Employee.java
@@ -18,11 +18,11 @@ public class Employee {
     @Indexed(unique = true)
     private String email;
     @Indexed(unique = true)
-    private Integer phone;
+    private Long phone;
     // private Type type; Para después
 
 
-    public Employee(String firstName, String lastName, Gender gender, String email, Integer phone) {
+    public Employee(String firstName, String lastName, Gender gender, String email, Long phone) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.gender = gender;
