+++ src/main/java/com/alexis/meetingback/models/Meeting.java
@@ -14,11 +14,11 @@ public class Meeting {
     @MongoId
     private String id;
     private LocalDateTime date;
-    private Integer cost;
-    private Integer hours;
+    private Long cost;
+    private Long hours;
     private Status status;
 
-    public Meeting(LocalDateTime date, Integer cost, Integer hours, Status status) {
+    public Meeting(LocalDateTime date, Long cost, Long hours, Status status) {
         this.date = date;
         this.cost = cost;
         this.hours = hours;
